+++ batchglm/api/models/numpy/glm_nb.py
@@ -1,4 +1,4 @@
-from batchglm.models.glm_nb import Model, Simulator
+from batchglm.models.glm_nb import Model
 from batchglm.train.numpy.glm_nb import Estimator
 
-__all__ = ["Model", "Simulator", "Estimator"]
+__all__ = ["Model", "Estimator"]
