+++ batchglm/models/base_glm/__init__.py
@@ -1,5 +1,4 @@
 from .estimator import _EstimatorGLM
 from .input import InputDataGLM
 from .model import _ModelGLM
-from .simulator import _SimulatorGLM
 from .utils import closedform_glm_mean, closedform_glm_scale, parse_design
