+++ batchglm/models/base_glm/external.py
@@ -1,2 +1,3 @@
 import batchglm.data as data_utils
+from batchglm import pkg_constants
 from batchglm.utils.linalg import groupwise_solve_lm
