+++ batchglm/models/glm_beta/__init__.py
@@ -1,3 +1,2 @@
 from .external import _EstimatorGLM
 from .model import Model
-from .simulator import Simulator
