+++ batchglm/models/glm_norm/__init__.py
@@ -1,3 +1,2 @@
 from .external import _EstimatorGLM
 from .model import Model
-from .simulator import Simulator
