+++ batchglm/models/glm_norm/external.py
@@ -1,4 +1,4 @@
 import batchglm.data as data_utils
 from batchglm import pkg_constants
-from batchglm.models.base_glm import _EstimatorGLM, _ModelGLM, _SimulatorGLM, closedform_glm_mean, closedform_glm_scale
+from batchglm.models.base_glm import _EstimatorGLM, _ModelGLM, closedform_glm_mean, closedform_glm_scale
 from batchglm.utils.linalg import groupwise_solve_lm
